+++ app/app.global.css
@@ -15,6 +15,7 @@
 @import './components/calibrationInputs/TextKeyboard.css';
 @import './components/evolverConfigs/evolver-configs.css';
 @import './components/python-shell/ScriptFinder.css';
+@import './components/ScriptEditor.css';
